pyproject.toml
[tool.poetry]
name = "the-internet"
version = "0.1.0"
description = ""
authors = ["dddpr <ddd.pronin@yandex.ru>"]
readme = "README.md"
packages = [{include = "the_internet"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.2.2"
selenium = "^4.8.3"
webdriver-manager = "^3.8.5"
allure-pytest = "^2.13.1"


[tool.poetry.group.dev.dependencies]
flakeheaven = "^3.2.1"
isort = "^5.12.0"
flake8 = "4.0.1"
flake8-commas = "^2.1.0"
flake8-cognitive-complexity = "^0.1.0"
flake8-annotations = "2.9.1"
flake8-pie = "^0.16.0"
flake8-simplify = "^0.20.0"
flake8-expression-complexity = "^0.0.11"
flake8-docstrings = "^1.7.0"
flake8-typing-imports = "1.2"
flake8-import-order = "^0.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flakeheaven]
baseline = ".flakeheaven_baseline"
exclude = [
    ".git",
    ".venv",
    "reports/*"
]
max-line-length = 99
max-complexity = 10
suppress-none-returning = true
show-source = true
# flake8-quotes
inline-quotes = "single"
multiline-quotes = "double"

[tool.flakeheaven.plugins]
flake8-import-order = ["+*"]
"flake8-*" = ["+*"]
pycodestyle = [
    "+*",
    "-W503",  # line break before binary operator
]
pyflakes = ["+*"] # это понадобилось для проверки 'Imported but unused'
"flake8-annotations" = [
    "+*",
    "-ANN101",  # Missing type annotation for self in method
    "-ANN102",  # Missing type annotation for cls in classmethod
    "-ANN204",  # Missing return type annotation for special method
]
"flake8-docstrings" = [
    "+*",
    "-D100" # docs for imports
]
pylint = [
    "+*",
#    "-W0511",
#    "-C0103", # Argument name doesn't conform to snake_case naming style
#    "-C0116", # Docstrings for public methods
#    "-C0115", # Missing class docstring
#    "-C0114", # Missing module docstring
#    "-R0903", # Too few public methods
]
